move callback interfaces to algorithm module
