+++ src/collections/circularbuffer.ts
@@ -7,6 +7,10 @@
 |----------------------------------------------------------------------------*/
 module phosphor.collections {
 
+import ICallback = algorithm.ICallback;
+import IPredicate = algorithm.IPredicate;
+
+
 /**
  * A circular buffer with a fixed maximum size.
  *
