+++ src/collections/queue.ts
@@ -7,6 +7,10 @@
 |----------------------------------------------------------------------------*/
 module phosphor.collections {
 
+import ICallback = algorithm.ICallback;
+import IPredicate = algorithm.IPredicate;
+
+
 /**
  * A canonical singly linked FIFO queue.
  */
